rearranging code and fixing tests
