+++ qpc/insights/utils.py
@@ -131,70 +131,3 @@ def format_upload_success(streamdata):
         return stream_list[2]
     except IndexError:
         return streamdata
-
-
-def verify_report_details(filename):
-    """
-    Verify that the report contents are a valid deployments report.
-
-    :returns boolean regarding report validity
-    """
-    valid = True
-    report_contents = extract_json_from_tarfile(filename)
-
-    # validate report_platform_id
-    report_platform_id = report_contents.get('report_platform_id')
-    if not report_platform_id:
-        valid = False
-        print(_(messages.INSIGHTS_REPORT_MISSING_FIELD % 'report_platform_id'))
-
-    # validate report id
-    report_id = report_contents.get('report_id')
-    if not report_id:
-        valid = False
-        print(_(messages.INSIGHTS_REPORT_MISSING_FIELD % 'report_id'))
-
-    # validate version type
-    report_version = report_contents.get('report_version')
-    if not report_version:
-        valid = False
-        print(_(messages.INSIGHTS_REPORT_MISSING_FIELD % 'report_version'))
-
-    # validate report type
-    report_type = report_contents.get('report_type', '')
-    if report_type != 'deployments':
-        valid = False
-        print(_(messages.INSIGHTS_MISSING_OR_INVALID % 'report_type'))
-
-    # validate system fingerprints
-    fingerprints = report_contents.get('system_fingerprints')
-    if not fingerprints:
-        valid = False
-        print((messages.INSIGHTS_REPORT_MISSING_FIELD % 'system_fingerprints'))
-
-    if fingerprints and report_platform_id:
-        verified_fingerprints = verify_report_fingerprints(fingerprints, report_id)
-        if not verified_fingerprints:
-            valid = False
-            print(_(messages.INSIGHTS_REPORT_NO_VALID_FP % report_id))
-    return valid
-
-
-def verify_report_fingerprints(fingerprints, report_id):
-    """Verify that report fingerprints contain canonical facts."""
-    canonical_facts = ['insights_client_id', 'bios_uuid', 'ip_addresses',
-                       'mac_addresses', 'vm_uuid', 'etc_machine_id',
-                       'subscription_manager_id']
-    verified_fingerprints = []
-    for fingerprint in fingerprints:
-        found_facts = False
-        for fact in canonical_facts:
-            if fingerprint.get(fact):
-                found_facts = True
-                break
-        if found_facts:
-            verified_fingerprints.append(fingerprint)
-        else:
-            print(_(messages.INSIGHTS_REPORT_INVALID_FP %
-                    (report_id, fingerprint.pop('metadata', None))))
-    return verified_fingerprints
